refactor: extract tool execution helpers from engine/loop_helpers.py
