+++ src/synthorg/engine/hybrid_loop.py
@@ -46,8 +46,6 @@ from .loop_helpers import (
     check_shutdown,
     check_stagnation,
     classify_turn,
-    clear_last_turn_tool_calls,
-    execute_tool_calls,
     get_tool_definitions,
     invoke_compaction,
     make_turn_record,
@@ -59,6 +57,10 @@ from .loop_protocol import (
     ShutdownChecker,
     TerminationReason,
     TurnRecord,
+)
+from .loop_tool_execution import (
+    clear_last_turn_tool_calls,
+    execute_tool_calls,
 )
 from .plan_helpers import update_step_status
 from .plan_models import (
